+++ src/view/film-card-view.js
@@ -1,13 +1,13 @@
 import {createElement} from '../render.js';
 
 const createNewFilmCard = (film) => {
-  const {title, poster, description} = film.filmInfo;
+  const {title, poster, description, totalRating} = film.filmInfo;
 
   return (
     `<article class="film-card">
       <a class="film-card__link">
         <h3 class="film-card__title">${title}</h3>
-        <p class="film-card__rating">8.3</p>
+        <p class="film-card__rating">${totalRating}</p>
         <p class="film-card__info">
           <span class="film-card__year">1929</span>
           <span class="film-card__duration">1h 55m</span>
